requirements.txt
# PyTorch dependencies are installed separately in Dockerfile for CUDA compatibility
# torch and torchvision are installed via: torch torchvision --index-url https://download.pytorch.org/whl/cu121

# Core dependencies for FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.8.6
aiofiles==23.2.1
gunicorn==21.2.0

# ML/AI dependencies for BiRefNet
transformers
timm
kornia
numpy
scipy==1.11.2
einops==0.8.1

# Image processing
opencv-python
Pillow==9.4.0
matplotlib
pycocotools==2.0.6

# Utilities
requests==2.31.0
tqdm
packaging
sentencepiece
shortuuid==1.0.11

# Optional/extra dependencies
markdown2==2.4.10
ftfy
hydra-core
accelerate

# Monitoring
prometheus-client==0.17.1

# Development/testing (optional)
pytest
mypy
